REFACTORED ENTIRE CODEBASE : ALMOST DONE
