+++ agri-store-react/src/Components/Others/Authentication.js
@@ -1,5 +1,4 @@
-import React from "react";
-import { Navigate, useLocation } from "react-router-dom";
+import { React, Navigate, useLocation } from "../../Utils/CustomUtils";
 
 function Authentication({ children }) {
   const location = useLocation();
