+++ agri-store-react/src/Components/ProductCard/ProductCard.js
@@ -1,5 +1,4 @@
-import React from "react";
-import { Link } from "react-router-dom";
+import { React, Link, toast } from "../../Utils/CustomUtils";
 import {
   useWishlistContext,
   useCartContext,
