+++ agri-store-react/src/Components/TotalPrice/TotalPrice.css
