+++ agri-store-react/src/Context/CartPageContext.js
@@ -1,4 +1,4 @@
-import React, { createContext, useContext } from "react";
+import { React, createContext, useContext } from "../Utils/CustomUtils";
 import { useProductContext } from "./ProductContext";
 const cartContext = createContext();
 export const useCartContext = () => useContext(cartContext);
