+++ agri-store-react/src/Services/WishlistServices.js
@@ -1,5 +1,4 @@
-import axios from "axios";
-import toast from "react-hot-toast";
+import { React, toast, axios } from "../Utils/CustomUtils";
 
 export const addToWishlist = async (productdata, dispatch) => {
   const response = await axios({
