+++ agri-store-react/src/Utils/CustomCSSUtils.css
@@ -1,0 +1,16 @@
+@import "../Components/Accountdetails/Accountdetails.css";
+@import "../Components/AccountTabs/AccountTabs.css";
+@import "../Components/AddAddress/Address.css";
+@import "../Components/Checkout/Checkout.css";
+@import "../Components/FeaturerdBrands/FeaturedBrands.css";
+@import "../Components/Footer/Footer.css";
+@import "../Components/Header/Header.css";
+@import "../Components/Hero/Hero.css";
+@import "../Components/LoadingSpinner/Spinner.css";
+@import "../Components/Login-Signup-Inputs/LoginInputs.css";
+@import "../Components/Login-Signup-Inputs/SignupInputs.css";
+@import "../Components/OrderConfirm/OrderConfirm.css";
+@import "../Components/ProductCard/ProductCard.css";
+@import "../Components/Sidebar/Sidebar.css";
+@import "../Components/SingleProductCard/SingleProductCard.css";
+@import "../Components/TotalPrice/TotalPrice.css";
